requirements.txt
telethon
nest-asyncio
python-dateutil
python-dotenv
pytest
pytest-asyncio
openai